+++ backend/src/main/java/dev/tripdraw/draw/application/RouteImageUploader.java
@@ -1,8 +1,8 @@
-package dev.tripdraw.application.draw;
+package dev.tripdraw.draw.application;
 
-import static dev.tripdraw.exception.draw.DrawExceptionType.IMAGE_SAVE_FAIL;
+import static dev.tripdraw.draw.exception.DrawExceptionType.IMAGE_SAVE_FAIL;
 
-import dev.tripdraw.exception.draw.DrawException;
+import dev.tripdraw.draw.exception.DrawException;
 import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
