+++ backend/src/main/java/dev/tripdraw/draw/exception/DrawException.java
@@ -1,4 +1,4 @@
-package dev.tripdraw.exception.draw;
+package dev.tripdraw.draw.exception;
 
 import dev.tripdraw.exception.common.BaseException;
 import dev.tripdraw.exception.common.ExceptionType;
