+++ backend/src/test/java/dev/tripdraw/draw/application/PostCreateEventHandlerTest.java
@@ -1,4 +1,4 @@
-package dev.tripdraw.application.draw;
+package dev.tripdraw.draw.application;
 
 import static dev.tripdraw.test.TestFixture.감상;
 import static dev.tripdraw.test.TestFixture.여행;
@@ -9,6 +9,8 @@ import static org.mockito.Mockito.times;
 
 import dev.tripdraw.domain.post.PostCreateEvent;
 import dev.tripdraw.domain.post.PostRepository;
+import dev.tripdraw.draw.application.PostCreateEventHandler;
+import dev.tripdraw.draw.application.RouteImageGenerator;
 import dev.tripdraw.trip.domain.TripRepository;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
