+++ backend/src/test/java/dev/tripdraw/draw/application/TripUpdateEventHandlerIntegrationTest.java
@@ -1,48 +1,53 @@
-package dev.tripdraw.application.draw;
+package dev.tripdraw.draw.application;
 
 import static dev.tripdraw.test.TestFixture.여행;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
-import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.timeout;
 
+import dev.tripdraw.draw.application.RouteImageGenerator;
 import dev.tripdraw.trip.domain.TripRepository;
 import dev.tripdraw.trip.domain.TripUpdateEvent;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.transaction.support.TransactionTemplate;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-@ExtendWith(MockitoExtension.class)
-class TripUpdateEventHandlerTest {
+@SpringBootTest
+public class TripUpdateEventHandlerIntegrationTest {
 
-    @Mock
+    @MockBean
     private RouteImageGenerator routeImageGenerator;
 
-    @Mock
+    @MockBean
     private TripRepository tripRepository;
 
-    @InjectMocks
-    private TripUpdateEventHandler tripUpdateEventHandler;
+    @Autowired
+    private ApplicationEventPublisher applicationEventPublisher;
+
+    @Autowired
+    private TransactionTemplate transactionTemplate;
 
     @Test
-    void 여행수정_이벤트를_받아_이미지를_생성_요청을_한다() {
+    void 여행수정_이벤트를_발생시키면_이미지를_생성_요청을_한다() {
         // given
         TripUpdateEvent tripUpdateEvent = new TripUpdateEvent(1L);
         given(tripRepository.getTripWithPoints(tripUpdateEvent.tripId()))
                 .willReturn(여행());
 
         // when
-        tripUpdateEventHandler.handle(tripUpdateEvent);
+        transactionTemplate.executeWithoutResult(action -> applicationEventPublisher.publishEvent(tripUpdateEvent));
 
         // then
         then(routeImageGenerator)
-                .should(times(1))
+                .should(timeout(5000).times(1))
                 .generate(any(), any(), any(), any());
     }
 }
