+++ backend/src/test/java/dev/tripdraw/draw/domain/PositionsTest.java
@@ -1,7 +1,9 @@
-package dev.tripdraw.domain.draw;
+package dev.tripdraw.draw.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import dev.tripdraw.draw.domain.Position;
+import dev.tripdraw.draw.domain.Positions;
 import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
